* fixing file import and export issues * fixing missing prop renaming from subtitle to subheading
