+++ packages/exhibit/src/common/ui/api/index.ts
@@ -1,1 +1,1 @@
-export { ConfiguredSystemHealthService, SystemHealthReport } from "./health";
+export { ConfiguredSystemHealthService, type SystemHealthReport } from "./health";
