+++ packages/features/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Adding BreadcrumbUtils for operating on breadcrumbs
 * Adding GalleryDirectory for displaying and selecting apps by family
 * Adding application-level Navigation with breadcrumbs
+* Exporting ConfiguredSendEmailHandler from `@otuekong-portfolio/archives`
 
 ### Changed
 * Project folder restructuring: moving all current contents into the `ui` folder
