+++ packages/features/src/galleries/server/index.ts
@@ -1,0 +1,1 @@
+export { ConfiguredSendEmailHandler } from "@otuekong-portfolio/archives";
