+++ packages/features/src/galleries/ui/helloworld/common/components/portfolioViewSection.tsx
@@ -10,7 +10,7 @@ import {
 } from "@otuekong-portfolio/archives";
 import { Section, SectionHandle } from "@otuekong-portfolio/curio";
 
-import { Breadcrumb } from "../../../../navigation";
+import { Breadcrumb } from "../../../../../navigation";
 
 import { PortfolioSubViewName } from "../portfolio.constants";
